Refactor: package relocations/renamings and unused classes removed
