--- app/common/application/controllers/error_responses/error_handlers.py
@@ -1,15 +1,0 @@
-from typing import Tuple
-
-from flask import jsonify, Response
-
-from app.common.application.controllers.response_status import ResponseStatus
-
-
-def render_bad_request_problem(error) -> Tuple[Response, int]:
-    response = {
-        "status": ResponseStatus.failure.value,
-        "status_code": None,
-        "message": error.detail
-    }
-
-    return jsonify(response), error.status
