--- app/common/application/controllers/responses/error/error_response.py
@@ -1,7 +1,0 @@
-from abc import ABC
-
-
-class ErrorResponse(ABC):
-    http_code: int
-    status_code: str
-    message: str
