--- app/common/application/controllers/responses/error/error_response_serializer.py
@@ -1,14 +1,0 @@
-from typing import Tuple, Dict
-
-from app.common.application.controllers.responses.error.error_response import ErrorResponse
-from app.common.application.controllers.responses.response_status import ResponseStatus
-
-
-class ErrorResponseSerializer:
-
-    def serialize(self, error_response: ErrorResponse) -> Tuple[Dict, int]:
-        return {
-                   "status": ResponseStatus.failure.value,
-                   "status_code": error_response.status_code,
-                   "message": error_response.message
-               }, error_response.http_code
