--- app/common/application/controllers/responses/error/transfer_ingest_error_response.py
@@ -1,9 +1,0 @@
-from app.common.application.controllers.responses.error.error_response import ErrorResponse
-
-
-class TransferIngestErrorResponse(ErrorResponse):
-    http_code = 500
-    status_code = "TRANSFER_INGEST_ERROR"
-
-    def __init__(self, message: str) -> None:
-        self.message = message
