+++ app/common/application/controllers/responses/error_handlers.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 from flask import jsonify, Response
 
-from app.common.application.controllers.responses.response_status import ResponseStatus
+from app.common.application.response_status import ResponseStatus
 
 
 def render_bad_request_problem(error) -> Tuple[Response, int]:
