+++ app/common/application/controllers/responses/transfer_ingest_error_response.py
@@ -1,4 +1,4 @@
-from app.common.application.controllers.error_responses.error_response import ErrorResponse
+from app.common.application.controllers.responses.error_response import ErrorResponse
 
 
 class TransferIngestErrorResponse(ErrorResponse):
