+++ app/common/auto_name_enum.py
