+++ app/containers.py
@@ -1,17 +1,10 @@
 from dependency_injector import containers, providers
 
-from app.common.application.controllers.responses.error.error_response_serializer import ErrorResponseSerializer
+from app.common.application.controllers.responses.error_response_serializer import ErrorResponseSerializer
 from app.ingest.application.mq.publishers.process_ready_queue_publisher import ProcessReadyQueuePublisher
 from app.ingest.application.mq.publishers.transfer_ready_queue_publisher import TransferReadyQueuePublisher
 from app.ingest.data.repositories.ingest_repository import IngestRepository
 from app.ingest.domain.services.ingest_service import IngestService
-from app.common.application.controllers.error_responses.error_response_serializer import ErrorResponseSerializer
-
-
-class Serializers(containers.DeclarativeContainer):
-    error_response_serializer = providers.Factory(
-        ErrorResponseSerializer
-    )
 
 
 class Controllers(containers.DeclarativeContainer):
