+++ app/drs_import_management_service_app.py
@@ -1,7 +1,7 @@
 from connexion import FlaskApp
 from connexion.exceptions import BadRequestProblem
 
-from app.common.application.controllers.responses.error.error_handlers import render_bad_request_problem
+from app.common.application.controllers.responses.error_handlers import render_bad_request_problem
 from app.common.application.middlewares.authorization_middleware import AuthorizationMiddleware
 from app.drs_import_management_service_resolver import DrsImportManagementServiceResolver
 from app.health.application.controllers.health_get_controller import HealthGetController
