+++ app/ingest/application/controllers/ingest_post_controller.py
@@ -2,10 +2,10 @@ from typing import Tuple, Dict
 
 from flask import request
 
-from app.common.application.controllers.responses.error.error_response_serializer import ErrorResponseSerializer
-from app.common.application.controllers.responses.error.transfer_ingest_error_response import \
+from app.common.application.controllers.responses.error_response_serializer import ErrorResponseSerializer
+from app.common.application.controllers.responses.transfer_ingest_error_response import \
     TransferIngestErrorResponse
-from app.common.application.controllers.responses.response_status import ResponseStatus
+from app.common.application.response_status import ResponseStatus
 from app.containers import Services, Controllers
 from app.ingest.domain.models.ingest.depositing_application import DepositingApplication
 from app.ingest.domain.models.ingest.ingest import Ingest
